Config release 1.0.2 (#68)
* Animate timer progress changes.
* Switch back to AlarmManager for short alarms.
* Restore AlarmManager's deleted timer observer.
* Injected AlarmScheduler properties can be private.
* Signal to observers immediately when cycle starts.
* Remove unnecessary Cycle.isFinishingPhase property.
* TimerPresenter properties can be private.
* Reorder TimerPresenter methods.
* Fire an intermediate event on timer start to prevent UI lag.
* Refactor PhaseProgress data class.
* Expose Phase progress Observable with a function instead of a property.
* Preserve paused timer mode across configuration changes.
* Update Gradle plugin from 2.2.0-beta3 to 2.2.0-rc1.
* Bump version number from 1.0.1 to 1.0.2

+++ app/src/main/java/com/itsronald/twenty2020/timer/TimerContract.kt
@@ -95,14 +95,14 @@ interface TimerContract {
          * @param progress The current progress value to display.
          * @param maxProgress The maximum progress value the can be displayed.
          */
-        fun showWorkProgress(progress: Int, maxProgress: Int)
+        fun showWorkProgress(progress: Float, maxProgress: Float)
 
         /**
          * Display break phase progress in the view.
          * @param progress The current progress value to display.
          * @param maxProgress The maximum progress value the can be displayed.
          */
-        fun showBreakProgress(progress: Int, maxProgress: Int)
+        fun showBreakProgress(progress: Float, maxProgress: Float)
 
         /**
          * Change the FloatingActionButton's drawable icon in the view.
